Relocating sciond api to scion_elem
